Add documentation and initial implementation of map/reduce worker.
Documentation describes information about the map / reduce process,
although it does not entirely match the implementation. This exists
more for reference purposes.

The map / reduce functionality is implemented in terms of a module
that is run in a different mode depending on arguments passed to it
by a scheduling process. It utilizes S3 for data storage, redis for
message passing. It assumes (for now) that data is sorted prior to
the map stage (allowing us to avoid sorting later). This implementation
is not particularly efficient, and future work can be done to enhance
functionality and speed.