Update formats, worker, and package.json
Describe JSON formats used in the worker / splitter configuration.
Update the worker to write locally to the filesystem; this will save
tons of potential memory issues (and save me from having to fork knox).
Update package.json to remove dependency on memorystream since we can
write to the filesystem in Heroku Cedar stack.